10.8+, preexisting node required.
I'm presently giving 10.7 support the old college try (building against the 10.7 SDK with Xcode 5.1, not necessarily supporting the Xcode 4.6 toolchain), but I thought this should be updated in the meantime. It also seems possible to me that the pre-installed node requirement is somewhat accidental. Since we have the source checked out already, it should be rather trivial to build an intermediate "bootstrap" node if there isn't one found in the path. Do it?