Allow unlimited concurrent connections
Some applications may need to be allowed to make more than the
hard-coded limit of 6 connections to a remote host. As an initial
patch, this removes the limite entirely. Though before this is
merged upstream I think we should add a JS API or something to
allow the caller to override limits on a per-host, or even per-call
basis.