Made the tests ignore the line ending
I had 4 asar tests fail because, the testing assumed that the file
content ends with UNIX line endings (\n), while it was using Windows
line endings (git probably changed them when I cloned the project on
Windows).
Since the line ending is not what matters in the tests, I trimmed them
and now all the tests pass as expected.