Link to open-file event in File object docs
Since we're on the subject of files it makes sense to include a link to info on the `open-file` event. Otherwise, it can only be found by reading about the `app` module for the main process which a user may not think to look at when trying to find documentation on interacting with files.