Pass correct window to menu items
In situation when there are two windows open, with one being in
focus and another one being blurred, if we popup menu for the blurred
window, click callbacks for menu items will get either `null` or focused
window as second parameter.
That fix passes information about window used in `popup` method with
fallback to a focused window.

Fix #8209