Make sure the LoadURLOptions interface is uniform
This was failing the build of electron-typescript-definitions
because it was unable to reuse a previously declared LoadURLOptions
interface due to the mismatch in types of the httpReferrer parameter.

I'm gonna need someone to verify that this extra type is actually
allowed for all the incantations of loadURL. Two of them I feel
confident about because the doc mentions one being the exact same
as the other.