feat: upgrade electron-download to 4.x
This is a breaking change, as it causes the `electron` module to require
Node 4.x (which has been EOL'd anyway).

Upgrading adds features such as a customizable cache location (via the
`ELECTRON_CACHE` environment variable).

Originally merged in https://github.com/electron/electron/pull/10922 but
was reverted because it was going into a version of Electron that did
not allow breaking changes.