fix: ExternalArrayBuffer can't be called by Blink
Since Electron is a combination of Chromium + Node, N-API native module
can be run in Browser(Blink) Process. The problemis Blink API (eg:
WebGLContext.texImage2D(ArrayBufferView) does not allow
an ExternalArrayBuffer if it is not created by Blink itself. This
patch allows N-API ExternalArrayBuffer can be used in Blink.