fix: allow pulling in a later version of Clang
Due to an MSVC interoperability defect in the version of Clang used for
Chromium 75, it's necessary to build with a later version of Clang. This
change allows the correct version to be specified by setting
LLVM_FORCE_REVISION=version in the environment during a gclient sync.
We will provide a documentation update when a fixed version of Clang is
available. This PR is only intended to be a temporary workaround and
will be removed when Electron's Chromium updates to a new version.