fix: reduce clipboard flakiness on Windows on Arm
Sometimes, the clipboard section of the Electron unit tests fails
because Chromium could not acquire the clipboard. This patch increases
the number of times that Chromium is allowed to try and acquire the
clipboard. No functional changes on any other system are expected (since
on those systems, the clipboard is acquired within the original limit).
No impact is expected in reasonable scenarios (i.e. with regular user
use).