Update setAspectRatio documentation
The documentation suggested that `setAspectRatio` `extraSize` took an array of numbers as input, when in fact it's an object with `width` and `height` properties.

Also removed the line below, since it's simply not true. The typescript definition only allows you to send in an object, not a number. Unless it suggests sending `{height: 0, width: 0}`, but in that case the documentation needs to reflect that.

`Calling this function with a value of 0 will remove any previously set aspect
ratios.`